refector: move hooks file
