+++ src/pages/community/PostDetail.jsx
@@ -3,9 +3,9 @@ import { useLocation, useNavigate } from "react-router-dom";
 import { useQuery } from "@tanstack/react-query";
 import parse from "html-react-parser";
 import { searchComments } from "@/domains/post/api/post";
-import { deleteCommentMutation, updateCommentMutation } from "./hooks/useComment";
-import { useGetUserId } from "./hooks/useGetUserId";
-import { useDeleteTargetPostMutation } from "@/pages/community/hooks/useDeletePost";
+import { deleteCommentMutation, updateCommentMutation } from "../../domains/post/hooks/useComment";
+import { useGetUserId } from "../../domains/post/hooks/useGetUserId";
+import { useDeleteTargetPostMutation } from "@/domains/post/hooks/useDeletePost";
 import { PageTop, PageWrapper } from "@/shared/components/layout/PageTopTitle";
 import { Button } from "@/shared/components/buttons/Button";
 import { CommentWriteForm } from "./components/CommentWriteForm";
