+++ src/pages/community/components/NoticeTable.jsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { useGetNoticeList } from "@/domains/post/hooks/useGetPostList";
-import { useHandleDeletePosts } from "../hooks/useHandleDeletePosts";
-import { useGetUserId } from "../hooks/useGetUserId";
-import { useViewCount } from "../hooks/useViewCount";
+import { useHandleDeletePosts } from "../../../domains/post/hooks/useHandleDeletePosts";
+import { useGetUserId } from "../../../domains/post/hooks/useGetUserId";
+import { useViewCount } from "../../../domains/post/hooks/useViewCount";
 import { IconBtn } from "@/shared/components/buttons/IconBtn";
 import CheckboxSmall from "@/shared/components/checkbox/CheckboxSmall";
 import { OnlyCheckBox } from "@/shared/components/checkbox/CheckboxLabel";
